Added GET Operator

+++ src/main/java/com/warframe/squad/entity/Operator.java
@@ -6,9 +6,9 @@ import lombok.Data;
 @Data
 @Builder
 public class Operator {
-  private Integer operator_id;
+  private Long operator_id;
   private String operator_name;
-  private Integer warframe_id;
+  private Long warframe_id;
   private FocusSchool focus_school;
 
 }
