+++ src/main/java/com/warframe/squad/entity/Warframe.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Warframe {
-  private Integer warframe_id;
-  private Integer operator_id;
+  private Long warframe_id;
+  private Long operator_id;
   private String warframe_name;
-  private Integer primary_weapon;
-  private Integer secondary_weapon;
-  private Integer melee_weapon;
+  private Long primary_weapon;
+  private Long secondary_weapon;
+  private Long melee_weapon;
 
 }
